Create python_music_player.py
This script when executed, in the same directory, as other mp3 files, choose one by random and plays it the song for the first 100 seconds due to the "time.sleep(100)" command that prevents the for-loop to auto run through them all. 

For future modifications or improvements, this script should extract the length of each music file and play the entire song then move to the next song.