Create using_random_to_flip_a_coin.js
In JavaScript, I wrote code that utilizes Math.random() to give a number between 1 - 1.99. If the number is 1.50 or greater, then the Math.round() is used on the number that is outputted from Math.random. An if and else statement is then used to print out, if the number is greater than or equal to 1.50, then it is Heads, otherwise, it is Tails.